+++ app/controllers/admin_controller.rb
@@ -5,9 +5,6 @@ class AdminController < ApplicationController
   before_action :authenticate_user!
   before_action :check_system_admin
 
-  def dashboard
-  end
-
   def roles
   end